Extract Nostr client initialization
